Fix certificate upload workflow and API proxy conflicts
- Fix 502 errors: run backend without --reload to prevent restarts on file changes
- Fix HTML-instead-of-JSON errors: normalize API paths with /api prefix in apiClient
- Move backend rewrites to beforeFileRewrites to prevent Next.js page route conflicts
- Add robust response parsing: read text first, then JSON.parse to avoid body stream issues
- Add lightweight /api/certificates/{id}/status endpoint for efficient frontend polling
- Add structured logging to certificate upload route
- Add polling timeout (40 attempts), failure detection (5 consecutive), and stage indicators
- Remove stale service worker files (sw.js, workbox) causing ChunkLoadError
- Clean up temporary test scripts

+++ database/db_setup.py
@@ -1,4 +1,4 @@
-from sqlalchemy import TIMESTAMP, Column, Date, Float, Index, Integer, String, func
+from sqlalchemy import TIMESTAMP, Column, Date, Index, Integer, Numeric, String, func
 
 from etl.logging_config import logger
 from shared.database import Base, get_engine
@@ -22,10 +22,10 @@ class SalesRecord(Base):
     category = Column(String(100), nullable=True, index=True)
     sub_category = Column(String(100), nullable=True)
     product_name = Column(String(500), nullable=True)
-    sales = Column(Float, nullable=False, default=0.0)
+    sales = Column(Numeric(18, 2), nullable=False, default=0.0)
     quantity = Column(Integer, nullable=False, default=0)
-    discount = Column(Float, nullable=False, default=0.0)
-    profit = Column(Float, nullable=False, default=0.0)
+    discount = Column(Numeric(18, 2), nullable=False, default=0.0)
+    profit = Column(Numeric(18, 2), nullable=False, default=0.0)
     created_at = Column(TIMESTAMP, server_default=func.now(), nullable=False)
     updated_at = Column(TIMESTAMP, server_default=func.now(), onupdate=func.now(), nullable=False)
